Move SkipOnEmpty callbacks to a separate folder
