+++ tests/ValidatorTest.php
@@ -18,9 +18,9 @@ use Yiisoft\Validator\Rule\Number;
 use Yiisoft\Validator\Rule\Required;
 use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\SimpleRuleHandlerContainer;
-use Yiisoft\Validator\SkipNever;
-use Yiisoft\Validator\SkipOnEmpty;
-use Yiisoft\Validator\SkipOnNull;
+use Yiisoft\Validator\SkipOnEmptyCallback\SkipNever;
+use Yiisoft\Validator\SkipOnEmptyCallback\SkipOnEmpty;
+use Yiisoft\Validator\SkipOnEmptyCallback\SkipOnNull;
 use Yiisoft\Validator\Tests\Stub\DataSet;
 use Yiisoft\Validator\Tests\Stub\FakeValidatorFactory;
 use Yiisoft\Validator\Tests\Stub\Rule;
